+++ tests/test_unit/test_arguments.py
@@ -198,42 +198,6 @@ def test_invalid_empty_exclude_argument(
     )
 
 
-@unix_only
-@parametrize_directives
-def test_invalid_empty_order_argument(
-    directive, page, tmp_path, caplog, plugin,
-):
-    directory_to_include = tmp_path / 'empty_order_argument'
-    directory_to_include.mkdir()
-
-    page_to_include_filepath = directory_to_include / 'included.md'
-    page_to_include_filepath.write_text('Content that should be included\n')
-
-    page_to_exclude_filepath = directory_to_include / 'igno"re"d.md'
-    page_to_exclude_filepath.write_text('Content that should be excluded\n')
-
-    includer_glob = os.path.join(str(directory_to_include), '*.md')
-
-    includer_file_content = f'''{{%
-  {directive} "{includer_glob}"
-  order=
-%}}'''
-
-    with pytest.raises(PluginError) as exc:
-        on_page_markdown(
-            includer_file_content,
-            page(tmp_path / 'includer.md'),
-            tmp_path,
-            plugin,
-        )
-
-    assert len(caplog.records) == 0
-    assert str(exc.value) == (
-        f"Invalid empty 'order' argument in '{directive}' directive"
-        ' at includer.md:1'
-    )
-
-
 @parametrize_directives
 def test_invalid_empty_encoding_argument(
     directive, page, tmp_path, plugin, caplog,
